Added location custom-post-type, and upon new location added, it's added to the post type as well. Delete works for that too. On the front end the url rewrite works. Created a template page on the plugin which loads if the theme doesn't have it

+++ src/Includes/ClassActivator.php
@@ -15,7 +15,7 @@ class ClassActivator {
     /**
      * This method gets fired during plugin activation.
      *
-     * @since       1.0.0
+     * @since       2020.11
      * @package     community-directory
      * @return      void
      */
@@ -60,7 +60,12 @@ class ClassActivator {
     }
 
     public static function install(){
-        community_directory_create_tables();
+        self::add_default_options();
+        
+        do_action( 'community_directory_create_tables' );
+
+        do_action( 'community_directory_flush_rewrite_rules' );
+        update_option('community_directory_flush_rewrite', 1);
 
         // update the version
         update_option('community_directory_db_version', COMMUNITY_DIRECTORY_DB_VERSION);
@@ -83,6 +88,28 @@ class ClassActivator {
         return $wpdb->get_col( $sql );
     }
 
+    /**
+     * Adds default settings during plugin activation.
+     *
+     * @since       2020.11
+     * @package     community-directory
+     * @return      void
+     */
+    public static function add_default_options() {
+        $settings = get_option( 'community_directory_settings', array());
+
+        $options = array(
+            'uninstall_erase_data' => 0,
+        );
+
+        foreach ($options as $option => $value){
+            if (!isset($settings[$option])) {
+                $settings[$option] = $value;
+            }
+        }
+
+        update_option( 'community_directory_settings', $settings );
+    }
 
     public static function deactivate($network_wide = false) {
         if (is_multisite()) {
@@ -123,12 +150,35 @@ class ClassActivator {
     }
 
     public static function uninstall() {
-        global $wpdb;
-        $tables = community_directory_drop_tables_on_delete_blog(array());
+        $options = get_option( 'community_directory_settings' );
+    
+        if ( isset($options['uninstall_erase_data']) &&
+            $options['uninstall_erase_data'] == '1'
+        ) {
+            global $wpdb;
 
-        $wpdb->query( "DROP TABLE IF EXISTS " . implode(',', $tables) );
+            $tables = ClassTables::drop_tables_on_delete_blog( array() );
+            foreach ( $tables as $table_name ) {
+                $sql        = "DROP TABLE IF EXISTS $table_name";
+                $wpdb->query( $sql );
+            }
 
-        delete_option('community_directory_db_version');
+            // Delete all location post type posts
+            $custom_post_type = ClassLocation::$location_post_type;
+            $sql = "DELETE FROM wp_posts WHERE post_type='$custom_post_type'";
+            $wpdb->query( $sql );
+            $sql = "DELETE FROM wp_postmeta WHERE post_id NOT IN (SELECT id FROM wp_posts)";
+            $wpdb->query( $sql );
+            $sql = "DELETE FROM wp_term_relationships WHERE object_id NOT IN (SELECT id FROM wp_posts)";
+            $wpdb->query( $sql );
+
+            // Delete options
+            delete_option( 'community_directory_settings' );
+            delete_option( 'community_directory_activation_redirect' );
+            delete_option( 'community_directory_flush_rewrite' );
+            delete_option( 'community_directory_default_data_installed' );
+            delete_option( 'community_directory_db_version' );
+        }
     }
